Исправление ошибок

+++ PsychologyApp.Application/Technique/ITechniqueService.cs
@@ -11,9 +11,9 @@ namespace PsychologyApp.Application.Services.TechniqueService;
 public interface ITechniqueService : IAppService
 {
     public Task<IList<TechniqueDTO>> GetTechniquesList(int count);
-    public Task<TechniqueDTO> GetTechniqueById(int id);
+    public Task<TechniqueDTO> GetTechniqueById(long id);
     public Task AddNewTechnique(TechniqueDTO techniqueDTO);
     public Task DeleteTechnique(TechniqueDTO techniqueDTO);
     public Task UpdateTechnique(TechniqueDTO techniqueDTO);
-    public Task MarkTechniqueAsCompleted(int id);
+    public Task MarkTechniqueAsCompleted(long id);
 }
