+++ PsychologyApp.Presentation/Views/TechniquePages/ConstructorPages/DesignerPage.xaml.cs
@@ -4,7 +4,7 @@ namespace MobileHelperMaui.Views.TechniquePages.ConstructorPages;
 
 public partial class DesignerPage : ContentPage
 {
-	public DesignerPage(int id)
+	public DesignerPage(long id)
 	{
 		InitializeComponent();
         this.BindingContext = new DesignerViewModel(this.Navigation, id);
